Rename bcifprint.py to BcifPrint.py
