+++ mmcif/tests/BcifPrint.py
